WIP init cli split btw extract and validate
